Controller export: Honor import status result
If the import exits unexpectedly, then do not provide the update files
as the new files for the agent.

+++ controller-export.sh
@@ -25,21 +25,24 @@ perform_export() {
 		echo "Skipping import of $project because export directory is empty"
 	else
 		listOfProjects="$project" gathererScripts="$gathererScripts" importerTasks="vcs,update,developerlink" logLevel="INFO" skipGather="true" IMPORTER_BASE="$CONTROLLER_DIRECTORY" SKIP_REQUIREMENTS="true" importerProperties="-Dimporter.relPath=$project/export" ./jenkins-scraper.sh
+		local status=$?
 
-		sudo chmod 2770 $agent_directory/update
-		sudo chmod 2770 $agent_directory/update/$project
-		sudo rm -rf $agent_directory/update/$project/*
-		for updateFile in $updateFiles; do
-			updatePath="$controller_directory/export/$project/$updateFile"
-			if [ -e "$updatePath" ]; then
-				cp $updatePath $agent_directory/update/$project/$updateFile
-			else
-				echo "Update file $updatePath could not be copied"
-			fi
-		done
-		sudo chown agent-$project:controller -R $agent_directory/update/$project
-		sudo chmod 2700 $agent_directory/update/$project
-		sudo chmod 2700 $agent_directory/update
+		if [ $status -eq 0 ]; then
+			sudo chmod 2770 $agent_directory/update
+			sudo chmod 2770 $agent_directory/update/$project
+			sudo rm -rf $agent_directory/update/$project/*
+			for updateFile in $updateFiles; do
+				updatePath="$controller_directory/export/$project/$updateFile"
+				if [ -e "$updatePath" ]; then
+					cp $updatePath $agent_directory/update/$project/$updateFile
+				else
+					echo "Update file $updatePath could not be copied"
+				fi
+			done
+			sudo chown agent-$project:controller -R $agent_directory/update/$project
+			sudo chmod 2700 $agent_directory/update/$project
+			sudo chmod 2700 $agent_directory/update
+		fi
 	fi
 
 	sudo chmod 2700 $agent_directory
